% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{about_server}
\alias{about_server}
\title{Shiny server for \emph{\strong{About}} page}
\usage{
about_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Output from \code{\link[shiny:moduleServer]{shiny::moduleServer()}}.
}
\description{
Shiny server for \emph{\strong{About}} page
}
\keyword{internal}
