% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{about_ui}
\alias{about_ui}
\title{Shiny UI for \emph{\strong{About}} page}
\usage{
about_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Output from \code{\link[shiny:reexports]{shiny::tagList()}}.
}
\description{
Shiny UI for \emph{\strong{About}} page
}
\keyword{internal}
