% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_citation.R
\name{citation_ui}
\alias{citation_ui}
\title{Shiny UI for \emph{\strong{Citation}} page}
\usage{
citation_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Output from \code{\link[shiny:reexports]{shiny::tagList()}}.
}
\description{
Shiny UI for \emph{\strong{Citation}} page
}
\keyword{internal}
