% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_tracing_input.R
\name{contact_tracing_input}
\alias{contact_tracing_input}
\title{Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising the
\code{symptomatic_traced} argument in \code{\link[ringbp:event_prob_opts]{ringbp::event_prob_opts()}}}
\usage{
contact_tracing_input(ns, ...)
}
\arguments{
\item{ns}{A namespace created with \code{\link[shiny:NS]{shiny::NS()}}.}

\item{...}{\link{dots} Not used, will throw a warning if arguments are supplied.}
}
\value{
A \code{\link[bslib:accordion]{bslib::accordion()}} object.
}
\description{
Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising the
\code{symptomatic_traced} argument in \code{\link[ringbp:event_prob_opts]{ringbp::event_prob_opts()}}
}
\keyword{internal}
