% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_tracing_input.R
\name{contact_tracing_seq_input}
\alias{contact_tracing_seq_input}
\title{Generate \code{\link[bslib:accordion]{bslib::accordion()}} with \code{From}/\code{To}/\code{By} numeric inputs that
parameterise a \code{\link[=seq]{seq()}} of contact tracing probabilities
(\code{symptomatic_traced} argument in \code{\link[ringbp:event_prob_opts]{ringbp::event_prob_opts()}}) to sweep over}
\usage{
contact_tracing_seq_input(ns, from, to, by, ...)
}
\arguments{
\item{ns}{A namespace created with \code{\link[shiny:NS]{shiny::NS()}}.}

\item{from, to, by}{\code{numeric} values input into \code{\link[=seq]{seq()}}.}

\item{...}{\link{dots} Not used, will throw a warning if arguments are supplied.}
}
\value{
A \code{\link[bslib:accordion]{bslib::accordion()}} object.
}
\description{
Generate \code{\link[bslib:accordion]{bslib::accordion()}} with \code{From}/\code{To}/\code{By} numeric inputs that
parameterise a \code{\link[=seq]{seq()}} of contact tracing probabilities
(\code{symptomatic_traced} argument in \code{\link[ringbp:event_prob_opts]{ringbp::event_prob_opts()}}) to sweep over
}
\keyword{internal}
