% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{delay_sampler}
\alias{delay_sampler}
\title{Build a random sampling function for a delay distribution}
\usage{
delay_sampler(distribution, par1, par2)
}
\arguments{
\item{distribution}{A \code{character} string, one of \code{"lnorm"}, \code{"gamma"} or
\code{"weibull"}.}

\item{par1, par2}{\code{numeric} scalars. \code{meanlog}/\code{sdlog} for the Lognormal,
\code{shape}/\code{scale} for the Gamma and Weibull.}
}
\value{
A \code{function} of a single argument \code{n}, as expected by
\code{\link[ringbp:delay_opts]{ringbp::delay_opts()}}.
}
\description{
Build a random sampling function for a delay distribution
}
\keyword{internal}
