% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_docs.R
\name{docs_link_card}
\alias{docs_link_card}
\title{Generate a card linking to one \pkg{ringbp} vignette}
\usage{
docs_link_card(title, description, href)
}
\arguments{
\item{title}{A \code{character} string naming the vignette.}

\item{description}{A \code{character} string summarising what the vignette covers.}

\item{href}{A \code{character} string with the URL of the vignette on the
\pkg{ringbp} package website.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} object.
}
\description{
Generate a card linking to one \pkg{ringbp} vignette
}
\keyword{internal}
