% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_docs.R
\name{docs_ui}
\alias{docs_ui}
\title{Shiny UI for \emph{\strong{Docs}} page}
\usage{
docs_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Output from \code{\link[shiny:tagList]{shiny::tagList()}}.
}
\description{
Shiny UI for \emph{\strong{Docs}} page
}
\details{
Links to the \pkg{ringbp} vignettes. Links open in a new tab, so that a
simulation in progress is not lost.

\pkg{ringbp} vignettes used to be embedded in \emph{\strong{propose}} using \code{iframe}s,
however there were issues with rendering these across the multiple
\emph{\strong{propose}} deployments.
}
\keyword{internal}
