% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_prob_input.R
\name{event_prob_input}
\alias{event_prob_input}
\title{Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising \code{asymptomatic},
\code{presymptomatic_transmission} and \code{symptomatic_ascertained} arguments in
\code{\link[ringbp:event_prob_opts]{ringbp::event_prob_opts()}}}
\usage{
event_prob_input(ns, ...)
}
\arguments{
\item{ns}{A namespace created with \code{\link[shiny:NS]{shiny::NS()}}.}

\item{...}{\link{dots} Not used, will throw a warning if arguments are supplied.}
}
\value{
A \code{\link[bslib:accordion]{bslib::accordion()}} object.
}
\description{
Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising \code{asymptomatic},
\code{presymptomatic_transmission} and \code{symptomatic_ascertained} arguments in
\code{\link[ringbp:event_prob_opts]{ringbp::event_prob_opts()}}
}
\keyword{internal}
