% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explore.R
\name{explore_server}
\alias{explore_server}
\title{Shiny server for \emph{\strong{Explore}} page}
\usage{
explore_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Output from \code{\link[shiny:moduleServer]{shiny::moduleServer()}}.
}
\description{
Shiny server for \emph{\strong{Explore}} page
}
\keyword{internal}
