% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incubation_input.R
\name{incubation_input}
\alias{incubation_input}
\title{Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising
\code{incubation_period} argument in \code{\link[ringbp:delay_opts]{ringbp::delay_opts()}}}
\usage{
incubation_input(ns, ...)
}
\arguments{
\item{ns}{A namespace created with \code{\link[shiny:NS]{shiny::NS()}}.}

\item{...}{\link{dots} Not used, will throw a warning if arguments are supplied.}
}
\value{
A \code{\link[bslib:accordion]{bslib::accordion()}} object.
}
\description{
Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising
\code{incubation_period} argument in \code{\link[ringbp:delay_opts]{ringbp::delay_opts()}}
}
\keyword{internal}
