% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_manual.R
\name{manual_ui}
\alias{manual_ui}
\title{Shiny UI for \emph{\strong{Manual}} page}
\usage{
manual_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Output from \code{\link[shiny:reexports]{shiny::tagList()}}.
}
\description{
Shiny UI for \emph{\strong{Manual}} page
}
\keyword{internal}
