% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_input.R
\name{npi_activation_day_feedback_server}
\alias{npi_activation_day_feedback_server}
\title{Register input-validation feedback for the \code{npi_activation_day} intervention
input rendered by \code{\link[=intervention_input]{intervention_input()}}}
\usage{
npi_activation_day_feedback_server(input)
}
\arguments{
\item{input}{The Shiny \code{input} reactive of the calling module.}
}
\value{
Invisible \code{NULL}; called for side-effects.
}
\description{
Register input-validation feedback for the \code{npi_activation_day} intervention
input rendered by \code{\link[=intervention_input]{intervention_input()}}
}
\keyword{internal}
