% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{offspring_sampler}
\alias{offspring_sampler}
\title{Build a random sampling function for an offspring distribution}
\usage{
offspring_sampler(distribution, r0, disp)
}
\arguments{
\item{distribution}{A \code{character} string, one of \code{"nbinom"}, \code{"pois"} or
\code{"geom"}.}

\item{r0, disp}{\code{numeric} scalars for the reproduction number and the
Negative Binomial dispersion.}
}
\value{
A \code{function} of a single argument \code{n}, as expected by
\code{\link[ringbp:offspring_opts]{ringbp::offspring_opts()}}.
}
\description{
Build a random sampling function for an offspring distribution
}
\keyword{internal}
