% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_outbreak_size.R
\name{outbreak_size_server}
\alias{outbreak_size_server}
\title{Shiny server for \emph{\strong{Outbreak Size & Length}} page}
\usage{
outbreak_size_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Output from \code{\link[shiny:moduleServer]{shiny::moduleServer()}}.
}
\description{
Shiny server for \emph{\strong{Outbreak Size & Length}} page
}
\keyword{internal}
