% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_outbreak_size.R
\name{outbreak_size_ui}
\alias{outbreak_size_ui}
\title{Shiny UI for \emph{\strong{Outbreak Size & Length}} page}
\usage{
outbreak_size_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Output from \code{\link[shiny:reexports]{shiny::tagList()}}.
}
\description{
Shiny UI for \emph{\strong{Outbreak Size & Length}} page
}
\keyword{internal}
