% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_title.R
\name{page_title}
\alias{page_title}
\title{Create title for \pkg{propose} page (i.e. \code{\link[bslib:nav-items]{bslib::nav_panel()}})}
\usage{
page_title(title, ...)
}
\arguments{
\item{title}{A \code{character} string of length 1.}

\item{...}{Extra arguments, not currently used (will warn if used).}
}
\value{
A \code{\link[shiny:reexports]{shiny::tags()}} \code{div}.
}
\description{
Create title for \pkg{propose} page (i.e. \code{\link[bslib:nav-items]{bslib::nav_panel()}})
}
\keyword{internal}
