% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patho_param_input.R
\name{patho_param_input}
\alias{patho_param_input}
\title{Generate \code{\link[bslib:card]{bslib::card()}} with a \code{\link[shiny:selectInput]{shiny::selectInput()}} for selecting the
pathogen parameters for the simulation}
\usage{
patho_param_input(ns, ...)
}
\arguments{
\item{ns}{A namespace created with \code{\link[shiny:NS]{shiny::NS()}}.}

\item{...}{\link{dots} Not used, will throw a warning if arguments are supplied.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} object.
}
\description{
Generate \code{\link[bslib:card]{bslib::card()}} with a \code{\link[shiny:selectInput]{shiny::selectInput()}} for selecting the
pathogen parameters for the simulation
}
\keyword{internal}
