% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reset_pathogen_params}
\alias{reset_pathogen_params}
\title{Reset the pathogen parameters in the \emph{\strong{Explore}} page to default values}
\usage{
reset_pathogen_params(session, defaults)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}

\item{defaults}{A list of pathogen parameters, one of the sublist from
\link{PROPOSE_DEFAULTS}.}
}
\value{
Nothing, called for side-effects from \verb{shiny::update*()} functions.
}
\description{
Reset the pathogen parameters in the \emph{\strong{Explore}} page to default values
}
\keyword{internal}
