% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_input.R
\name{sim_feedback_server}
\alias{sim_feedback_server}
\title{Register input-validation feedback for \code{\link[=sim_input]{sim_input()}}}
\usage{
sim_feedback_server(input)
}
\arguments{
\item{input}{The Shiny \code{input} reactive of the calling module.}
}
\value{
Invisible \code{NULL}; called for side-effects.
}
\description{
Register input-validation feedback for \code{\link[=sim_input]{sim_input()}}
}
\keyword{internal}
