% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_input.R
\name{sim_input}
\alias{sim_input}
\title{Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising the simulation
controls.}
\usage{
sim_input(ns, ...)
}
\arguments{
\item{ns}{A namespace created with \code{\link[shiny:NS]{shiny::NS()}}.}

\item{...}{\link{dots} Not used, will throw a warning if arguments are supplied.}
}
\value{
A \code{\link[bslib:accordion]{bslib::accordion()}} object.
}
\description{
Including parameterising:
\itemize{
\item \code{cap_max_days} and \code{cap_cases} arguments in
\code{\link[ringbp:sim_opts]{ringbp::sim_opts()}}
\item The seed used in the simulation (see \code{\link[=set.seed]{set.seed()}}). By default the seed
is set to \code{NA} which is interpreted in the server as choosing a random
seed for each time the simulation is run. A single \code{numeric} can be set
to specify the seed.
}
}
\keyword{internal}
