% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptom_event_prob_input.R
\name{symptom_event_prob_feedback_server}
\alias{symptom_event_prob_feedback_server}
\title{Register input-validation feedback for \code{\link[=symptom_event_prob_input]{symptom_event_prob_input()}}}
\usage{
symptom_event_prob_feedback_server(input)
}
\arguments{
\item{input}{The Shiny \code{input} reactive of the calling module.}
}
\value{
Invisible \code{NULL}; called for side-effects.
}
\description{
Register input-validation feedback for \code{\link[=symptom_event_prob_input]{symptom_event_prob_input()}}
}
\keyword{internal}
