% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptom_event_prob_input.R
\name{symptom_event_prob_input}
\alias{symptom_event_prob_input}
\title{Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising \code{asymptomatic}
and \code{presymptomatic_transmission} arguments in \code{\link[ringbp:event_prob_opts]{ringbp::event_prob_opts()}}}
\usage{
symptom_event_prob_input(ns, ...)
}
\arguments{
\item{ns}{A namespace created with \code{\link[shiny:NS]{shiny::NS()}}.}

\item{...}{\link{dots} Not used, will throw a warning if arguments are supplied.}
}
\value{
A \code{\link[bslib:accordion]{bslib::accordion()}} object.
}
\description{
Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising \code{asymptomatic}
and \code{presymptomatic_transmission} arguments in \code{\link[ringbp:event_prob_opts]{ringbp::event_prob_opts()}}
}
\keyword{internal}
