% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tooltip}
\alias{tooltip}
\alias{shape_tip}
\alias{scale_tip}
\title{Generate \code{\link[bslib:tooltip]{bslib::tooltip()}} message for delay distribution parameters}
\usage{
shape_tip(distribution, epiparameter)

scale_tip(distribution, epiparameter)
}
\arguments{
\item{distribution}{a \code{character} string for the distribution type, e.g.,
\code{"Gamma"} or \code{"Weibull"}.}

\item{epiparameter}{a \code{character} string for the name of the epidemiological
parameter, e.g. \code{"incubation period"} or \code{"onset-to-isolation delay"}.}
}
\value{
a \code{character} string
}
\description{
Generate \code{\link[bslib:tooltip]{bslib::tooltip()}} message for delay distribution parameters
}
\keyword{internal}
