% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tracing_effectiveness.R
\name{tracing_effectiveness_ui}
\alias{tracing_effectiveness_ui}
\title{Shiny UI for \emph{\strong{Contact Tracing Effectiveness}} page}
\usage{
tracing_effectiveness_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Output from \code{\link[shiny:tagList]{shiny::tagList()}}.
}
\description{
Shiny UI for \emph{\strong{Contact Tracing Effectiveness}} page
}
\keyword{internal}
