% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tracing_strategies.R
\name{tracing_strategies_server}
\alias{tracing_strategies_server}
\title{Shiny server for \emph{\strong{Contact Tracing Strategies}} page}
\usage{
tracing_strategies_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Output from \code{\link[shiny:moduleServer]{shiny::moduleServer()}}.
}
\description{
Shiny server for \emph{\strong{Contact Tracing Strategies}} page
}
\keyword{internal}
