% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tracing_strategies.R
\name{tracing_strategies_ui}
\alias{tracing_strategies_ui}
\title{Shiny UI for \emph{\strong{Contact Tracing Strategies}} page}
\usage{
tracing_strategies_ui(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
Output from \code{\link[shiny:reexports]{shiny::tagList()}}.
}
\description{
Shiny UI for \emph{\strong{Contact Tracing Strategies}} page
}
\keyword{internal}
