% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_tracing_input.R
\name{contact_tracing_feedback_server}
\alias{contact_tracing_feedback_server}
\title{Register input-validation feedback for \code{\link[=contact_tracing_input]{contact_tracing_input()}}}
\usage{
contact_tracing_feedback_server(input)
}
\arguments{
\item{input}{The Shiny \code{input} reactive of the calling module.}
}
\value{
Invisible \code{NULL}; called for side-effects.
}
\description{
Register input-validation feedback for \code{\link[=contact_tracing_input]{contact_tracing_input()}}
}
\keyword{internal}
