% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{defaults}
\alias{defaults}
\alias{PROPOSE_DEFAULTS}
\title{\pkg{ringbp} simulation defaults used in \pkg{propose}}
\format{
An object of class \code{list} of length 12.
}
\usage{
PROPOSE_DEFAULTS
}
\description{
\code{PROPOSE_DEFAULTS} is a list of lists and values. Each inner list contains
default pathogen parameters for the outbreak simulation for a specific
pathogen. The non-list values in the \code{PROPOSE_DEFAULTS} are the non-pathogen
parameters for the outbreak simulation.
}
\keyword{datasets}
