% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{defaults}
\alias{defaults}
\alias{PROPOSE_DEFAULTS}
\title{\pkg{ringbp} simulation defaults used in \pkg{propose}}
\format{
An object of class \code{list} of length 20.
}
\usage{
PROPOSE_DEFAULTS
}
\description{
\code{PROPOSE_DEFAULTS} is a list with the default outbreak simulation parameters.
}
\keyword{datasets}
