% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delays_input.R
\name{delays_input}
\alias{delays_input}
\title{Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising
\code{incubation_period} and \code{onset_to_isolation} arguments in
\code{\link[ringbp:delay_opts]{ringbp::delay_opts()}}}
\usage{
delays_input(ns, defaults = PROPOSE_DEFAULTS, ...)
}
\arguments{
\item{ns}{A namespace created with \code{\link[shiny:NS]{shiny::NS()}}.}

\item{...}{\link{dots} Not used, will throw a warning if arguments are supplied.}
}
\value{
A \code{\link[bslib:accordion]{bslib::accordion()}} object.
}
\description{
Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising
\code{incubation_period} and \code{onset_to_isolation} arguments in
\code{\link[ringbp:delay_opts]{ringbp::delay_opts()}}
}
\keyword{internal}
