% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_docs.R
\name{docs_ui}
\alias{docs_ui}
\title{Shiny UI for \emph{\strong{Docs}} page}
\usage{
docs_ui(id, vignette_file_name)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{vignette_file_name}{A \code{character} string with the file name of the
\pkg{ringbp} vignette, including the \code{.html} file extension.}
}
\value{
Output from \code{\link[shiny:reexports]{shiny::tagList()}}.
}
\description{
Shiny UI for \emph{\strong{Docs}} page
}
\keyword{internal}
