% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_input.R
\name{intervention_input}
\alias{intervention_input}
\title{Generate \code{\link[bslib:accordion]{bslib::accordion()}} with input for parameterising the
\code{quarantine} argument in \code{\link[ringbp:intervention_opts]{ringbp::intervention_opts()}}}
\usage{
intervention_input(ns, ...)
}
\arguments{
\item{ns}{A namespace created with \code{\link[shiny:NS]{shiny::NS()}}.}

\item{...}{\link{dots} Not used, will throw a warning if arguments are supplied.}
}
\value{
A \code{\link[bslib:accordion]{bslib::accordion()}} object.
}
\description{
Generate \code{\link[bslib:accordion]{bslib::accordion()}} with input for parameterising the
\code{quarantine} argument in \code{\link[ringbp:intervention_opts]{ringbp::intervention_opts()}}
}
\keyword{internal}
