% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offspring_input.R
\name{offspring_input}
\alias{offspring_input}
\title{Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising \code{community} and
\code{isolated} arguments in \code{\link[ringbp:offspring_opts]{ringbp::offspring_opts()}}}
\usage{
offspring_input(ns, ...)
}
\arguments{
\item{ns}{A namespace created with \code{\link[shiny:NS]{shiny::NS()}}.}

\item{...}{\link{dots} Not used, will throw a warning if arguments are supplied.}
}
\value{
A \code{\link[bslib:accordion]{bslib::accordion()}} object.
}
\description{
Generate \code{\link[bslib:accordion]{bslib::accordion()}} with inputs for parameterising \code{community} and
\code{isolated} arguments in \code{\link[ringbp:offspring_opts]{ringbp::offspring_opts()}}
}
\keyword{internal}
