% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_input.R
\name{replicates_input}
\alias{replicates_input}
\title{Generate \code{\link[bslib:card]{bslib::card()}} with input for number of simulation replicates}
\usage{
replicates_input(ns, ...)
}
\arguments{
\item{ns}{A namespace created with \code{\link[shiny:NS]{shiny::NS()}}.}

\item{...}{\link{dots} Not used, will throw a warning if arguments are supplied.}
}
\value{
A \code{\link[bslib:card]{bslib::card()}} object.
}
\description{
Generate \code{\link[bslib:card]{bslib::card()}} with input for number of simulation replicates
}
\keyword{internal}
